--- src/st/collect_dummy.h
@@ -1,2 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-static void CollectDummy(u16 id) { DestroyEntity(g_CurrentEntity); }
